backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "google-adk>=1.25.0",
    "google-genai>=1.68.0",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "typer>=0.24.1",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.8",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
